rock paper scissors game
created a working rock paper scissors game within javascript.

initially I was stuck on this a few weeks ago, but after teaching
myself the fundamentals of psuedocode and how  to
split the problem down further, I spent a few hours writing the
psuedocode then converted to comments
I then fleshed the comments out with the javascript syntax and
started to test each stage using console.log() to ensure it worked.